backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.11.4,<3.0
pydantic-core>=2.11.4
pydantic[email]
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.8.0 
beanie==1.29.0
motor>=3.1.1,<4.0
pymongo==4.10.1
email-validator
openai>=1.11.0
pytest==7.4.0
httpx==0.25.0
pytest-asyncio==0.21.1